LayoutTests/fast/media/mq-hover-cssom-expected.txt

Test the CSSOM serialization of the hover media feature.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: none)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: none)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(hover: ON-DEMAND)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(Hover: On-demand)"
PASS successfullyParsed is true

TEST COMPLETE
